matroska/ParserUtils.java:
skip() calls InputStream.skip(), which is allowed to return 0, for example
if reading from a named pipe or fifo file that's blocked or trying to
read past the end of the file. If skip isn't working, try reading one
byte to either unstick it or at least advance the file pointer so it
doesn't end up with an infinite loop.

An alternative would be to just give up if input.skip() returns 0 because
I'm not sure recovery is possible.