MP4Reader:
There's a tiny window if during a thread pool/executor is being shutdown
because the process is exiting or something, trying to acquire the lock
might throw a InterruptedException exception.  If that occurs, don't try to
release the lock we didn't get or we'll crash here with a less helpful
exception.